Bump to 0.1.5

+++ src/crc.rs
@@ -41,7 +41,7 @@ impl<R: Reader> CrcReader<R> {
 }
 
 impl<R: Reader> Reader for CrcReader<R> {
-    fn read(&mut self, into: &mut [u8]) -> IoResult<uint> {
+    fn read(&mut self, into: &mut [u8]) -> IoResult<usize> {
         let amt = try!(self.inner.read(into));
         self.crc.update(into.slice_to(amt));
         Ok(amt)
